# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/orders": {
      "post": {
        "description": "Single-cook order. Stock is reserved for `available` products.\n\n**Delivery (required)** — copied onto the order at create time (profile changes later do not affect this order):\n\n- `source: CUSTOMER_PROFILE` — uses `PATCH /users/me/delivery-information`\n- `source: CUSTOM` — requires `location` (GeoJSON Point `[lng, lat]`) and `address`; `additionalInformation` optional",
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              },
              "examples": {
                "customerProfile": {
                  "summary": "Use saved customer delivery",
                  "value": {
                    "cookId": "507f1f77bcf86cd799439011",
                    "items": [
                      {
                        "productId": "507f1f77bcf86cd799439012",
                        "quantity": 2
                      }
                    ],
                    "delivery": {
                      "source": "CUSTOMER_PROFILE"
                    },
                    "paymentMethod": "nequi"
                  }
                },
                "customDelivery": {
                  "summary": "Custom address for this order",
                  "value": {
                    "cookId": "507f1f77bcf86cd799439011",
                    "items": [
                      {
                        "productId": "507f1f77bcf86cd799439012",
                        "quantity": 2
                      }
                    ],
                    "delivery": {
                      "source": "CUSTOM",
                      "location": {
                        "type": "Point",
                        "coordinates": [
                          -77.0319,
                          3.8833
                        ]
                      },
                      "address": "Calle 5 #10-20",
                      "additionalInformation": "Casa azul, next to the bakery"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created order including delivery snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create order with delivery snapshot",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "description": "**Customers (spec 003)** — order history scoped to the JWT user id (never pass `customerId`).\nSupports cursor pagination: `?limit=20&cursor=...` (default 20, max 100).\n\nResponse for customers:\n`{ data: [{ id, status, paymentStatus, total, createdAt }], pagination: { nextCursor, hasMore } }`\n\n**Cooks / admins** — same pagination shape; cooks only see orders for their cook profile (full order objects in `data`).",
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size (1–100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from the previous page `pagination.nextCursor`",
            "schema": {
              "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated order list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOrdersListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List orders for the authenticated user",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order by id",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Advance order status (cook/admin) or cancel PENDING/CONFIRMED (customer/cook/admin)",
        "tags": [
          "orders"
        ]
      }
    },
    "/api/orders/{id}/payment": {
      "patch": {
        "operationId": "OrdersController_updatePayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update payment (customer reports paid; cook marks PAID)",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/orders/{id}/confirm-reception": {
      "post": {
        "operationId": "OrdersController_confirmReception",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Customer confirms order reception",
        "tags": [
          "orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Petronio API",
    "description": "Marketplace API for cooks and customers (Pacific region).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DeliveryGeoPointDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "Point"
            ],
            "example": "Point"
          },
          "coordinates": {
            "example": [
              -77.0319,
              3.8833
            ],
            "description": "[longitude (-180..180), latitude (-90..90)]",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "coordinates"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "object",
            "nullable": true,
            "example": "eyJjcmVhdGVkQXQiOiIyMDI2LTA4LTE2VDEyOjAwOjAwLjAwMFoiLCJpZCI6IjY4YWYifQ"
          },
          "hasMore": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "hasMore"
        ]
      },
...
      "CreateOrderItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateOrderDeliveryDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "CUSTOMER_PROFILE",
              "CUSTOM"
            ],
            "example": "CUSTOMER_PROFILE",
            "description": "CUSTOMER_PROFILE copies saved user delivery; CUSTOM requires location + address"
          },
          "location": {
            "description": "Required when source is CUSTOM. GeoJSON Point [longitude, latitude]",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryGeoPointDto"
              }
            ]
          },
          "address": {
            "type": "string",
            "example": "Calle 5 #10-20",
            "description": "Required when source is CUSTOM (human-readable address)"
          },
          "additionalInformation": {
            "type": "string",
            "example": "Casa azul, next to the bakery",
            "description": "Optional notes; for CUSTOM only (profile path uses saved notes)"
          }
        },
        "required": [
          "source"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "cookId": {
            "type": "string",
            "description": "Single-cook order (MVP)"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderItemDto"
            }
          },
          "delivery": {
            "$ref": "#/components/schemas/CreateOrderDeliveryDto"
          },
          "shippingCost": {
            "type": "number",
            "example": 10000,
            "default": 0
          },
          "paymentMethod": {
            "type": "string",
            "example": "nequi"
          }
        },
        "required": [
          "cookId",
          "items",
          "delivery"
        ]
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439012"
          },
          "name": {
            "type": "string",
            "example": "Encocado de camarón"
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "unitPrice": {
            "type": "number",
            "example": 35000
          },
          "total": {
            "type": "number",
            "example": 70000
          }
        },
        "required": [
          "productId",
          "name",
          "quantity",
          "unitPrice",
          "total"
        ]
      },
      "OrderTotalsResponseDto": {
        "type": "object",
        "properties": {
          "subtotal": {
            "type": "number",
            "example": 70000
          },
          "shipping": {
            "type": "number",
            "example": 10000
          },
          "total": {
            "type": "number",
            "example": 80000
          }
        },
        "required": [
          "subtotal",
          "shipping",
          "total"
        ]
      },
      "OrderDeliveryResponseDto": {
        "type": "object",
        "properties": {
          "location": {
            "description": "Snapshot of delivery coordinates at order creation",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeliveryGeoPointDto"
              }
            ]
          },
          "address": {
            "type": "string",
            "example": "Calle 5 #10-20"
          },
          "additionalInformation": {
            "type": "string",
            "example": "Casa azul, next to the bakery",
            "nullable": true
          }
        },
        "required": [
          "location",
          "address"
        ]
      },
      "OrderPaymentResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAYMENT_INSTRUCTIONS",
              "PAID",
              "FAILED",
              "CANCELLED"
            ],
            "example": "PENDING"
          },
          "method": {
            "type": "string",
            "example": "nequi"
          },
          "customerReportedPaid": {
            "type": "boolean",
            "example": false
          },
          "paidAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-15T12:00:00.000Z"
          }
        },
        "required": [
          "status",
          "customerReportedPaid"
        ]
      },
      "OrderShippingResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PREPARING",
              "READY_TO_SHIP",
              "SHIPPED",
              "IN_TRANSIT",
              "DELIVERED"
            ],
            "example": "PENDING"
          },
          "carrier": {
            "type": "string",
            "example": "Servientrega"
          },
          "trackingNumber": {
            "type": "string",
            "example": "123456789"
          },
          "shippedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deliveredAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "CustomerConfirmationResponseDto": {
        "type": "object",
        "properties": {
          "confirmed": {
            "type": "boolean",
            "example": false
          },
          "confirmedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "confirmed"
        ]
      },
      "OrderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "68af1a2b3c4d5e6f78901234"
          },
          "orderNumber": {
            "type": "string",
            "example": "ORDER-000001"
          },
          "customerId": {
            "type": "string"
          },
          "cookId": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "totals": {
            "$ref": "#/components/schemas/OrderTotalsResponseDto"
          },
          "delivery": {
            "description": "Delivery address/location copied at create time; not linked to profile updates",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderDeliveryResponseDto"
              }
            ]
          },
          "payment": {
            "$ref": "#/components/schemas/OrderPaymentResponseDto"
          },
          "shipping": {
            "$ref": "#/components/schemas/OrderShippingResponseDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY_TO_SHIP",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ],
            "example": "PENDING"
          },
          "customerConfirmation": {
            "$ref": "#/components/schemas/CustomerConfirmationResponseDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-15T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-15T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "customerId",
          "cookId",
          "items",
          "totals",
          "delivery",
          "payment",
          "shipping",
          "status",
          "customerConfirmation",
          "createdAt",
          "updatedAt"
        ]
      },
      "CustomerOrderHistoryItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "68af1a2b3c4d5e6f78901234"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY_TO_SHIP",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ],
            "example": "SHIPPED"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAYMENT_INSTRUCTIONS",
              "PAID",
              "FAILED",
              "CANCELLED"
            ],
            "example": "PAID"
          },
          "total": {
            "type": "number",
            "example": 75000
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-08-15T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "status",
          "paymentStatus",
          "total",
          "createdAt"
        ]
      },
      "CustomerOrdersListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerOrderHistoryItemDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PREPARING",
              "READY_TO_SHIP",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdatePaymentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAYMENT_INSTRUCTIONS",
              "PAID",
              "FAILED",
              "CANCELLED"
            ]
          },
          "method": {
            "type": "string"
          },
          "customerReportedPaid": {
            "type": "boolean",
            "description": "Customer reports external payment was made"
          }
        }
      },
...
    }
  }
}